backend /requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==20.3.0
bech32==1.2.0
betterproto==2.0.0b5
bip32utils==0.3.post4
boltons==21.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cohere==5.13.12
cryptography==44.0.0
ecdsa==0.16.1
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
furl==2.1.3
greenlet==3.1.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.1
hyperframe==6.0.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.4
miscreant==0.3.0
mnemonic==0.19
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.0
ollama==0.4.4
orderedmultidict==1.0.1
orjson==3.10.12
packaging==24.2
propcache==0.2.1
protobuf==3.20.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
secp256k1==0.14.0
secret-ai-sdk==0.1.3
secret-sdk==1.8.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.45.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wheel==0.44.0
wrapt==1.17.0
yarl==1.18.3
